requirements-docker.txt
# ===========================================
# Core LlamaIndex
# ===========================================
llama-index
llama-index-core

# ===========================================
# Document Processing - Docling
# ===========================================
llama-index-readers-docling
docling

# ===========================================
# Ollama Integration (Open Source LLM)
# ===========================================
llama-index-embeddings-ollama
llama-index-llms-ollama

# ===========================================
# OpenAI Integration (Optional)
# ===========================================
llama-index-embeddings-openai
llama-index-llms-openai

# ===========================================
# Vector Store - PostgreSQL + pgvector
# ===========================================
llama-index-vector-stores-postgres
llama-index-storage-docstore-postgres
psycopg2-binary==2.9.9

# ===========================================
# Search - Hybrid (Semantic + BM25)
# ===========================================
rank_bm25
llama-index-retrievers-bm25

# ===========================================
# CrewAI - Multi-Agent Orchestration
# ===========================================
crewai
crewai-tools

# ===========================================
# RAGAs - Evaluation Framework
# ===========================================
ragas
datasets

# ===========================================
# Arize Phoenix - Observability
# ===========================================
arize-phoenix
arize-phoenix-otel
openinference-instrumentation-crewai
openinference-instrumentation-litellm

# ===========================================
# Web Framework
# ===========================================
fastapi
uvicorn
python-dotenv

# ===========================================
# Utilities
# ===========================================
sentencepiece
numpy<2
